refactor(provider): extract provider definitions into provider/definitions.ts
Move the provider definition domain out of extension.ts: PROVIDERS table,
ProviderDefinition / OpenCodeModel / ModelListEntry / ModelListResponse /
ConvertedMessageResult / LanguageModelConfiguration types, plus the
getUserAgent / isTransientFetchError helpers.

extension.ts now imports from provider/definitions. Behavior-preserving;
compile + 291 tests + lint green.
