+++ src/extension.ts
@@ -60,8 +60,6 @@ import {
   DEFAULT_STREAM_IDLE_TIMEOUT_SECONDS,
   DEFAULT_VISION_PROXY_PROMPT,
   EXTENSION_ID,
-  FALLBACK_USER_AGENT,
-  FREE_ZEN_MODEL_IDS,
   IMAGE_TOKEN_ESTIMATE,
   KNOWN_UNAVAILABLE_MODEL_IDS,
   MAX_HISTORY_IMAGES_KEPT,
@@ -81,7 +79,6 @@ import {
   SETTING_AUTO_ENABLE_AGENTS_WINDOW,
   SETTING_DEBUG_REASONING,
   SETTING_ENABLED,
-  SETTING_FREE_ONLY,
   SETTING_MAX_INPUT_TOKENS,
   SETTING_MAX_TOKENS,
   SETTING_REQUEST_TIMEOUT_SECONDS,
@@ -161,275 +158,19 @@ import {
   usageTrackerOptions,
 } from "./usage/dashboard";
 import { clearOpenCodeModelMetadataCache, getModelMetadataSnapshot, getOpenCodeModelMetadata } from "./models/metadataFetcher";
-
-interface ProviderDefinition {
-  vendor: AllProviderVendor;
-  displayName: string;
-  modelNamePrefix: string;
-  modelsUrl: string;
-  chatCompletionsUrl: string;
-  messagesUrl: string;
-  responsesUrl?: string;
-  testModelId: string;
-  fallbackModels: string[];
-  filterModel?: (modelId: string) => boolean;
-  /** When true, this provider only serves agent-host models (targetChatSessionType=copilotcli). */
-  isAgentVariant?: boolean;
-  /** The vendor key for the main (non-agent) provider definition this variant mirrors. */
-  baseVendor?: typeof GO_VENDOR | typeof ZEN_VENDOR;
-}
-
-type ModelEndpointKind = "chat-completions" | "messages" | "responses" | "google";
-
-let cachedUserAgent: string | undefined;
-
-/**
- * Build the User-Agent string from the extension's declared version.
- *
- * CONTRACT:
- * - Reads `context.extension.packageJSON.version` once, caches the result.
- * - Falls back to {@link FALLBACK_USER_AGENT} when version is unavailable
- *   (e.g. tests that construct a stub context).
- * - Avoids the drift that previously hardcoded a version literal here
- *   (issue #78: header reported `0.3.6` while package.json was `0.4.1`).
- */
-function getUserAgent(): string {
-  if (cachedUserAgent) return cachedUserAgent;
-  const packageJSON = vscode.extensions.getExtension("ltmoerdani.opencode-copilot-chat")?.packageJSON as { version?: unknown } | undefined;
-  const version = typeof packageJSON?.version === "string" ? packageJSON.version : undefined;
-  cachedUserAgent = version ? `opencode-copilot-chat/${version} VSCode` : FALLBACK_USER_AGENT;
-  return cachedUserAgent;
-}
-
-/**
- * Classify a fetch error as transient (worth retrying) vs. permanent.
- *
- * RULES:
- * - Network-layer errors (DNS, TCP reset, connect timeout, socket errors)
- *   are transient — undici exposes the real code via `error.cause`.
- * - HTTP 4xx (except 408/429) is permanent — retrying won't help.
- * - HTTP 408/429/5xx is transient — gateway/rate-limit style failures.
- *   These arrive via the "Model list request failed (NNN): ..." message
- *   that `fetchModels()` throws on a non-2xx response.
- * - AbortError from a CancellationToken is NEVER retried. TimeoutError from
- *   AbortSignal.timeout is transient and can be retried.
- */
-function isTransientFetchError(error: unknown): boolean {
-  // DOMException is a global since Node 17; guard anyway so a hypothetical
-  // older host never crashes inside error classification.
-  if (typeof DOMException === "function" && error instanceof DOMException) {
-    if (error.name === "AbortError") return false;
-    if (error.name === "TimeoutError") return true;
-  }
-  const cause = (error as { cause?: { code?: string; name?: string } } | undefined)?.cause;
-  const code = cause?.code ?? (error as { code?: string } | undefined)?.code;
-  const name = cause?.name ?? (error as { name?: string } | undefined)?.name;
-  // undici network error codes
-  if (code && /^E(AI_AGAIN|CONNRESET|CONNREFUSED|CONNABORTED|TIMEDOUT|HOSTUNREACH|NETUNREACH|PROTO|PIPE)$/.test(code)) {
-    return true;
-  }
-  if (name && /^UND_ERR_(CONNECT_TIMEOUT|SOCKET|REQUEST_TIMEOUT)$/.test(name)) {
-    return true;
-  }
-  // TypeError: fetch failed (the generic wrapper undici throws) — always retry;
-  // if the cause turns out to be non-transient, the inner check above handles it.
-  if (error instanceof TypeError && /fetch failed/i.test(error.message)) return true;
-  // Extract HTTP status from either an explicit `.status` field or the
-  // "Model list request failed (NNN): ..." message pattern.
-  const explicitStatus = (error as { status?: number } | undefined)?.status;
-  const msg = getErrorMessage(error);
-  const msgMatch = msg.match(/\((\d{3})\)/);
-  const httpStatus = typeof explicitStatus === "number" ? explicitStatus : msgMatch ? Number(msgMatch[1]) : undefined;
-  if (typeof httpStatus === "number") {
-    if (httpStatus === 408 || httpStatus === 429 || httpStatus >= 500) return true;
-    return false;
-  }
-  return false;
-}
-
-/** Create an agent-variant provider definition that inherits URLs, models, and filters from a base. */
-function providerVariant(
-  base: ProviderDefinition,
-  agentVendor: typeof AGENT_GO_VENDOR | typeof AGENT_ZEN_VENDOR,
-  displayName: string,
-): ProviderDefinition {
-  return {
-    vendor: agentVendor,
-    displayName,
-    modelNamePrefix: base.modelNamePrefix,
-    modelsUrl: base.modelsUrl,
-    chatCompletionsUrl: base.chatCompletionsUrl,
-    messagesUrl: base.messagesUrl,
-    responsesUrl: base.responsesUrl,
-    testModelId: base.testModelId,
-    fallbackModels: base.fallbackModels,
-    filterModel: base.filterModel,
-  };
-}
-
-const PROVIDERS: Record<ProviderDefinition["vendor"], ProviderDefinition> = (() => {
-  const go: ProviderDefinition = {
-    vendor: GO_VENDOR,
-    displayName: "OpenCode Go",
-    modelNamePrefix: "OpenCode Go",
-    modelsUrl: "https://opencode.ai/zen/go/v1/models",
-    chatCompletionsUrl: "https://opencode.ai/zen/go/v1/chat/completions",
-    messagesUrl: "https://opencode.ai/zen/go/v1/messages",
-    responsesUrl: "https://opencode.ai/zen/go/v1/responses",
-    testModelId: "deepseek-v4-flash",
-    fallbackModels: [
-      "deepseek-v4-pro",
-      "deepseek-v4-flash",
-      "glm-5.1",
-      "glm-5",
-      "hy3-preview",
-      "kimi-k2.6",
-      "kimi-k2.5",
-      "mimo-v2-omni",
-      "mimo-v2-pro",
-      "mimo-v2.5",
-      "mimo-v2.5-pro",
-      "minimax-m2.7",
-      "minimax-m2.5",
-      "qwen3.7-max",
-      "qwen3.7-plus",
-      "qwen3.6-plus",
-      "qwen3.5-plus",
-      "gpt-5.6-luna",
-    ],
-  };
-  const zen: ProviderDefinition = {
-    vendor: ZEN_VENDOR,
-    displayName: "OpenCode Zen",
-    modelNamePrefix: "OpenCode Zen",
-    modelsUrl: "https://opencode.ai/zen/v1/models",
-    chatCompletionsUrl: "https://opencode.ai/zen/v1/chat/completions",
-    messagesUrl: "https://opencode.ai/zen/v1/messages",
-    responsesUrl: "https://opencode.ai/zen/v1/responses",
-    testModelId: "deepseek-v4-flash-free",
-    fallbackModels: [
-      "claude-opus-4-7",
-      "claude-opus-4-6",
-      "claude-opus-4-5",
-      "claude-opus-4-1",
-      "claude-sonnet-4-6",
-      "claude-sonnet-4-5",
-      "claude-sonnet-4",
-      "claude-haiku-4-5",
-      "deepseek-v4-flash-free",
-      "gemini-3.5-flash",
-      "gemini-3.1-pro",
-      "gemini-3-flash",
-      "glm-5.1",
-      "glm-5",
-      "gpt-5.5",
-      "gpt-5.5-pro",
-      "gpt-5.4",
-      "gpt-5.4-pro",
-      "gpt-5.4-mini",
-      "gpt-5.4-nano",
-      "gpt-5.3-codex",
-      "gpt-5.3-codex-spark",
-      "gpt-5.2",
-      "gpt-5.2-codex",
-      "gpt-5.1",
-      "gpt-5.1-codex",
-      "gpt-5.1-codex-max",
-      "gpt-5.1-codex-mini",
-      "gpt-5",
-      "gpt-5-codex",
-      "gpt-5-nano",
-      "grok-build-0.1",
-      "kimi-k2.6",
-      "kimi-k2.5",
-      "minimax-m2.7",
-      "minimax-m2.5",
-      "minimax-m2.5-free",
-      "nemotron-3-super-free",
-      "qwen3.6-plus",
-      "qwen3.6-plus-free",
-      "qwen3.5-plus",
-      "big-pickle",
-    ],
-    filterModel: (modelId) =>
-      vscode.workspace.getConfiguration(CONFIG_SECTION).get<boolean>(SETTING_FREE_ONLY, true)
-        ? modelId.endsWith("-free") || FREE_ZEN_MODEL_IDS.has(modelId)
-        : true,
-  };
-  return {
-    [GO_VENDOR]: go,
-    [ZEN_VENDOR]: zen,
-    [AGENT_GO_VENDOR]: { ...providerVariant(go, AGENT_GO_VENDOR, "OpenCode Go (Agents)"), isAgentVariant: true, baseVendor: GO_VENDOR },
-    [AGENT_ZEN_VENDOR]: {
-      ...providerVariant(zen, AGENT_ZEN_VENDOR, "OpenCode Zen (Agents)"),
-      isAgentVariant: true,
-      baseVendor: ZEN_VENDOR,
-    },
-  };
-})();
-
-interface OpenCodeModel extends vscode.LanguageModelChatInformation {
-  endpointKind: ModelEndpointKind;
-  provider: ProviderDefinition;
-  rawModelId?: string;
-  isUserSelectable?: boolean;
-  configurationSchema?: vscode.LanguageModelConfigurationSchema;
-}
-
-interface ModelListEntry {
-  id?: string;
-  owned_by?: string;
-  status?: string;
-  deprecated?: boolean;
-  limit?: {
-    context?: number;
-    output?: number;
-  };
-  context_window?: number;
-  contextWindow?: number;
-  max_output_tokens?: number;
-  maxOutputTokens?: number;
-  attachment?: boolean;
-  image_input?: boolean;
-  imageInput?: boolean;
-  reasoning?: boolean;
-  modalities?: {
-    input?: string[];
-    output?: string[];
-  };
-}
-
-interface ModelListResponse {
-  data?: ModelListEntry[];
-}
-
-interface ConvertedMessageResult {
-  messages: ApiMessage[];
-  normalizedImageCount: number;
-}
-
-/**
- * Reasoning effort levels per model family, sourced from the upstream
- * OpenCode provider transform (anomalyco/opencode, packages/opencode/src/provider/transform.ts):
- *
- *   WIDELY_SUPPORTED_EFFORTS = ["low", "medium", "high"]
- *   OPENAI_EFFORTS = ["none", "minimal", "low", "medium", "high", "xhigh"]
- *
- * For @ai-sdk/openai-compatible (Mimo, and most models routed through
- * chat-completions): the default is WIDELY_SUPPORTED_EFFORTS = ["low", "medium", "high"].
- * DeepSeek V4 on openai-compatible additionally adds "max" → ["low", "medium", "high", "max"].
- */
-interface LanguageModelConfiguration {
-  apiKey?: unknown;
-}
-
-type ConfiguredLanguageModelInfoOptions = vscode.PrepareLanguageModelChatModelOptions & {
-  configuration?: LanguageModelConfiguration;
-};
-
-type ConfiguredLanguageModelResponseOptions = vscode.ProvideLanguageModelChatResponseOptions & {
-  configuration?: LanguageModelConfiguration;
-};
+import {
+  ConfiguredLanguageModelInfoOptions,
+  ConfiguredLanguageModelResponseOptions,
+  ConvertedMessageResult,
+  LanguageModelConfiguration,
+  ModelListEntry,
+  ModelListResponse,
+  OpenCodeModel,
+  PROVIDERS,
+  ProviderDefinition,
+  getUserAgent,
+  isTransientFetchError,
+} from "./provider/definitions";
 
 /**
  * Hard upper limit (in bytes of raw image data) for a single image embedded
